% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_classifications.R
\docType{data}
\name{available_classifications}
\alias{available_classifications}
\title{Available ocupational classifications}
\format{
A data frame with 6 rows and 3 variables:
\describe{
  \item{classification}{classification acronym}
  \item{classification_fullname}{classification full name in its own language}
  \item{country}{country of origin of ocupational classification}
  ...
}
}
\usage{
available_classifications
}
\description{
A dataset containing the names of available ocupational classifications in this package.
}
\keyword{datasets}
