% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2010_to_isco08.R
\name{census2010_to_isco08}
\alias{census2010_to_isco08}
\title{Cross Census 2010 (USA) to ISCO 08}
\usage{
census2010_to_isco08(base, census, summary = FALSE, code_titles = FALSE)
}
\arguments{
\item{base}{Dataframe from any survey including a variable with Standard Occupational Codes (SOC) 2010}

\item{census}{character vector containing Census 2010 codes}

\item{summary}{If TRUE provides other dataframe counting how many cases where asigned for each Census 08 code to each ISCO88 code.}

\item{code_titles}{If TRUE adds classification titles besides from codes.}
}
\value{
The function returns the provided dataframe, adding a new variable with the crosswalk to ISCO 08 codes
}
\description{
Cross Census 2010 (USA) to ISCO 08
}
\examples{

base_crossed <- census2010_to_isco08(toy_base_ipums_cps_2018,census = OCC2010,summary = TRUE)

}
