% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census2010_to_soc2010.R
\name{census2010_to_soc2010}
\alias{census2010_to_soc2010}
\title{Cross Census 2010 to SOC 2010}
\usage{
census2010_to_soc2010(base, census, code_titles = FALSE)
}
\arguments{
\item{base}{Dataframe from any Survey including a variable with USA Census 2010 codes}

\item{census}{character vector containing Census 2010 codes}

\item{code_titles}{logical, if TRUE code_titles are added to dataframe}
}
\value{
The function returns the provided dataframe, adding a new variable with the crosswalk to SOC 2010 codes
}
\description{
Cross Census 2010 to SOC 2010
}
\examples{

base_crossed <- census2010_to_soc2010(toy_base_ipums_cps_2018,census = "OCC2010",code_titles = FALSE)

}
