% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cno2017_to_isco_08.R
\name{cno2017_to_isco08}
\alias{cno2017_to_isco08}
\title{Cross CNO 2017 (Arg) to ISCO 08}
\usage{
cno2017_to_isco08(base, cno, summary = FALSE, code_titles = FALSE)
}
\arguments{
\item{base}{Dataframe from argentinian survey/census including a variable with CNO 2017 (Clasificador Nacional de Ocupaciones 2017) codes}

\item{cno}{character vector containing CNO 2017 codes}

\item{summary}{If TRUE provides other dataframe counting how many cases where asigned for each Census 08 code to each ISCO88 code.}

\item{code_titles}{If TRUE adds classification titles besides from codes.}
}
\value{
The function returns the provided dataframe, adding a new variable with the crosswalk to ISCO 08 codes
}
\description{
Cross CNO 2017 (Arg) to ISCO 08
}
\examples{

}
