% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_level.R
\name{complexity_level}
\alias{complexity_level}
\title{Occupations complexity level}
\usage{
complexity_level(base, isco, iscotype, summary = FALSE)
}
\arguments{
\item{base}{Dataframe from any survey including a variable with occupation codes}

\item{isco}{Character vector containing occupation codes}

\item{iscotype}{Type of occupational classificaction. ISCO08, ISCO88, Census2010, SOC2010 and CNO2017 are supported.}

\item{summary}{If TRUE provides other dataframe counting how many cases where asigned for each occupation code to each complexity level.}
}
\value{
The function returns the provided dataframe, adding a new variable with the occupation complexity level. The new complexity_level variable is a factor variable containg the levels 'Low', 'Medium' and 'High'.
}
\description{
Occupations complexity level
}
\examples{

base_complexity <- complexity_level(toy_base_ipums_cps_2018, OCC, iscotype="Census2010")
load("data/toy_base_uruguay.rda")
base_complexity <- complexity_level(toy_base_uruguay, f71_2, iscotype="ISCO08")

}
