% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstable_isco08_isco88.R
\docType{data}
\name{crosstable_isco08_isco88}
\alias{crosstable_isco08_isco88}
\title{Crosswalk between isco08 and isco88 classification.}
\format{
A data frame with 671 rows and 6 variables:
\describe{
  \item{ISCO-88 Title EN}{ISCO88 title}
  \item{ISCO-88 code}{ISCO88 code}
  \item{ISCO 08 Code}{ISCO08 code}
  \item{ISCO-08 part}{variable where 'p' indicates that there is more than one ISCO88 code for the ISCO08 provided}
  \item{ISCO-08 Title EN}{ISCO08 title}
  \item{Comments}{Other comments on the crosswalk}
  ...
}
}
\source{
\url{https://www.ilo.org/public/english/bureau/stat/isco/isco08/}
}
\usage{
crosstable_isco08_isco88
}
\description{
A dataset containing the crosswalk from isco08 to isco88 provided by
International Labour Organization.
}
\keyword{datasets}
