% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco_08_to_isco_88_4digit.R
\name{isco08_to_isco88_4digit}
\alias{isco08_to_isco88_4digit}
\title{Cross ISCO 08 a ISCO 88 - 4 digit}
\usage{
isco08_to_isco88_4digit(base, isco, summary = FALSE)
}
\arguments{
\item{base}{Dataframe from any survey including a variable with ISCO 08 codes}

\item{isco}{character vector containing ISCO 08 codes}

\item{summary}{If TRUE provides other dataframe counting how many cases where asigned for each ISCO08 code to each ISCO88 code.}
}
\value{
The function returns the provided dataframe, adding a new variable with the crosswalk to ISCO 88 codes
}
\description{
Cross ISCO 08 a ISCO 88 - 4 digit
}
\examples{

base_crossed <- isco08_to_isco88_4digit(toy_base_piaac,isco = "ISCO08_C",summary = TRUE)

}
