% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco_88_to_isco_08_3digit-Prueba.R
\name{isco88_to_isco08_3digit}
\alias{isco88_to_isco08_3digit}
\title{Cross ISCO 88 a ISCO 08 - 3 digit}
\usage{
isco88_to_isco08_3digit(base, isco, summary = FALSE)
}
\arguments{
\item{base}{Dataframe from any survey including a variable with ISCO 88 codes}

\item{isco}{character vector containing ISCO 88 codes at 3 digits}

\item{summary}{If TRUE provides other dataframe counting how many cases where asigned for each ISCO08 code to each ISCO88 code.}
}
\value{
The function returns the provided dataframe, adding a new variable with the crosswalk to ISCO 08 codes
}
\description{
Cross ISCO 88 a ISCO 08 - 3 digit
}
\examples{



}
