% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupationcross.R
\docType{package}
\name{ocucpationcross}
\alias{ocucpationcross}
\title{\code{ocucpationcross} package}
\description{
A package to easily apply crosswalks between occupational classifications such as
International Standard Classification of Occupations (ISCO) and Standard Occupational Classification (SOC)
}
\details{
See the README on
\href{https://github.com/Guidowe/occupationcross}{Github}
}
