% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify_to_isco08.R
\name{reclassify_to_isco08}
\alias{reclassify_to_isco08}
\title{Reclassify ocupational codes to International Standard Classification of Occupations 08}
\usage{
reclassify_to_isco08(
  base,
  variable,
  classif_origin,
  add_skill = F,
  code_titles = F,
  summary = F
)
}
\arguments{
\item{base}{Dataframe including a variable with occupational codes contained in "available_classifications"}

\item{variable}{variable containing occupational codes}

\item{classif_origin}{character vector specifying the Classification. "Census2010" (United States), "CNO2017" (Argentina), "SINCO2011" (Mexico) are supported.}

\item{add_skill}{If TRUE adds a new variable with the occupation skill level based on ISCO 08 skill levels. The new skill_level variable is a factor variable containg the levels 'Low', 'Medium' and 'High'.}

\item{code_titles}{If TRUE adds classification titles besides from codes.}

\item{summary}{If TRUE provides other dataframe counting how many cases where asigned for each Census 08 code to each ISCO88 code.}
}
\value{
The function returns the provided dataframe, adding a new variable with the crosswalk to ISCO 08 codes
}
\description{
Reclassify ocupational codes to International Standard Classification of Occupations 08
}
\details{
disclaimer: This script uses crosswalks provided by different statistical offices arround the world.
 It is not an official product of any of them
}
\examples{



USA_database_with_isco08 <- reclassify_to_isco08(toy_base_ipums_cps_2018, OCC, classif_origin="Census2010")
MEX_database_with_isco08 <- reclassify_to_isco08(toy_base_mexico, p3, classif_origin="SINCO2011")
ARG_database_with_isco08 <- reclassify_to_isco08(toy_base_eph_argentina, PP04D_COD, classif_origin="CNO2001")
}
