% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinco2011_to_isco_08.R
\name{sinco2011_to_isco08}
\alias{sinco2011_to_isco08}
\title{Cross SINCO 2011 (Mexico) to ISCO 08}
\usage{
sinco2011_to_isco08(base, sinco, summary = FALSE, code_titles = FALSE)
}
\arguments{
\item{base}{Dataframe from mexican survey including a variable with "Sistema Nacional de Clasificacion de Ocupaciones 2011" (SINCO 2011) codes}

\item{sinco}{character vector containing SINCO 2011 codes}

\item{summary}{If TRUE provides other dataframe counting how many cases where asigned for each Census 08 code to each ISCO88 code.}

\item{code_titles}{If TRUE adds classification titles besides from codes.}
}
\value{
The function returns the provided dataframe, adding a new variable with the crosswalk to ISCO 08 codes
}
\description{
Cross SINCO 2011 (Mexico) to ISCO 08
}
\examples{

base_crossed <- sinco2011_to_isco08(
 base = toy_base_mexico,
 sinco = p3,
 summary = FALSE,
 code_titles = TRUE)
}
