% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skill_level.R
\name{skill_level}
\alias{skill_level}
\title{Occupations skill level}
\usage{
skill_level(base, variable, classif_origin)
}
\arguments{
\item{base}{Dataframe from any survey including a variable with occupation codes}

\item{variable}{Character vector containing occupation codes}

\item{classif_origin}{Type of occupational classificaction. ISCO08, ISCO88, Census2010, SOC2010 and CNO2017 are supported.}
}
\value{
The function returns the provided dataframe, adding a new variable with the occupation skill level. The new skill_level variable is a factor variable containg the levels 'Low', 'Medium' and 'High'.
}
\description{
Occupations skill level
}
\examples{

base_uruguay_skill <- skill_level(toy_base_uruguay, f71_2, classif_origin="ISCO08")
base_mexico_skill <- skill_level(toy_base_mexico, p3, classif_origin = "SINCO2011")
base_eph_skill <- skill_level(toy_base_eph_argentina, PP04D_COD, classif_origin = "CNO2001")
base_cps_skill <- skill_level(toy_base_ipums_cps_2018, OCC2010, classif_origin =  "Census2010")


}
