% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prueba.R
\name{suma}
\alias{suma}
\title{Suma de vector, mas 2}
\usage{
suma(vector)
}
\arguments{
\item{vector}{Un vector numerico}
}
\value{
La suma de todos sus numeros mas 2
}
\description{
Suma de vector, mas 2
}
\examples{
suma(1:18)
}
