% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumatidy.R
\name{sumatidy}
\alias{sumatidy}
\title{Make a sorted frequency table for a factor}
\usage{
sumatidy(base)
}
\arguments{
\item{x}{factor}
}
\value{
A tibble
}
\description{
Make a sorted frequency table for a factor
}
\examples{


}
