% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_base_lfs.R
\docType{data}
\name{toy_base_lfs}
\alias{toy_base_lfs}
\title{Random sample of cases from Italy 2014 database provided by the Labour Force Survey.}
\format{
A data frame with 2000 rows and 11 variables:
\describe{
  \item{YEAR}{Survey year}
  \item{SEX}{Sex}
  \item{AGE}{Age}
  \item{WSTATOR}{COMPLETAR}
  \item{ILOSTAT}{COMPLETAR}
  \item{COUNTRYB}{COMPLETAR}
  \item{ISCO3D}{COMPLETAR}
  \item{ISCO1D}{COMPLETAR}
  \item{COEFF}{COMPLETAR}
  \item{HAT11LEV}{COMPLETAR}
  \item{SIZEFIRM}{COMPLETAR}
  ...
}
}
\usage{
toy_base_lfs
}
\description{
A dataset containing random cases from Italy 2014 LFS database, with a
selection of 11 variables, including ISCO codes.
}
\keyword{datasets}
