% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cno70_to_isco_08.R
\name{cno70_to_isco08}
\alias{cno70_to_isco08}
\title{Cross CNO-SENA 1970 (Colombia) to ISCO 08}
\usage{
cno70_to_isco08(base, oficio, summary = FALSE, code_titles = FALSE)
}
\arguments{
\item{base}{Dataframe from a colombian survey (e.g. GEIH) including a variable with CNO-SENA 1970 codes, at the 2-digit subgroup level published in the variable OFICIO}

\item{oficio}{character vector containing CNO-70 (OFICIO) 2-digit codes}

\item{summary}{If TRUE provides other dataframe counting how many cases where asigned for each origin code to each ISCO 08 code.}

\item{code_titles}{If TRUE adds classification titles besides from codes.}
}
\value{
The function returns the provided dataframe, adding a new variable with the crosswalk to ISCO 08 codes
}
\description{
Cross CNO-SENA 1970 (Colombia) to ISCO 08
}
\examples{

base_crossed <- cno70_to_isco08(toy_base_colombia_cno70, oficio = OFICIO)
}
