% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCO88_to_ISCO_08.R
\name{isco08_to_isco88}
\alias{isco08_to_isco88}
\title{Cross ISCO 08 a ISCO 88}
\usage{
isco08_to_isco88(base, isco, summary = F)
}
\arguments{
\item{base}{Dataframe from survey which includes ISCO 08 codes}

\item{isco}{Vector with isco 08 codes}
}
\value{
Dataframe adding a new variable with crosswalk to ISCO88
}
\description{
Cross ISCO 08 a ISCO 88
}
