% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISCO88_to_ISCO_08.R
\name{isco08_to_isco88}
\alias{isco08_to_isco88}
\title{Cross ISCO 08 a ISCO 88}
\usage{
isco08_to_isco88(base, isco, summary = FALSE)
}
\arguments{
\item{base}{Dataframe from any survey including a variable with ISCO 08 codes}

\item{isco}{character vector containing ISCO 08 codes}
}
\value{
The function returns the provided dataframe, adding a new variable with the crosswalk to ISCO 88 codes
}
\description{
Cross ISCO 08 a ISCO 88
}
\examples{

base_crossed <- isco08_to_isco88(toy_base_lfs,isco = "ISCO3D",summary = TRUE)

}
