% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco_88_to_isco_08_Ndigit.R
\name{isco88_to_isco08_ndigit}
\alias{isco88_to_isco08_ndigit}
\title{Cross ISCO 88 to ISCO 08 - n digit}
\usage{
isco88_to_isco08_ndigit(
  base,
  isco,
  digits,
  summary = FALSE,
  code_titles = FALSE
)
}
\arguments{
\item{base}{Dataframe from any survey including a variable with ISCO 08 codes}

\item{isco}{character vector containing ISCO 08 codes}

\item{digits}{Digits number (ranging from 1 to 4) for ISCO codes in input data}

\item{summary}{If TRUE provides other dataframe counting how many cases where asigned for each ISCO08 code to each ISCO88 code.}
}
\value{
The function returns the provided dataframe, adding a new variable with the crosswalk to ISCO 88 codes
}
\description{
Cross ISCO 88 to ISCO 08 - n digit
}
\examples{


toy_base_peru_isco08 <- isco88_to_isco08_ndigit(toy_base_peru,isco = p505,digits = 3)
}
